Added additional testing and fixed a bug.
