+++ src/test/java/language/LanguageParserTest.java
@@ -2,68 +2,55 @@ package language;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 import org.parboiled.Parboiled;
-import org.parboiled.Rule;
 import org.parboiled.parserunners.BasicParseRunner;
 import org.parboiled.support.ParsingResult;
 
-import java.util.Arrays;
-import java.util.Collection;
-
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
 
 /**
- * This class contains the tests for the LanguageParser.
+ * Contains the more complicated tests for LanguageParser.
  *
  * Created by Boudewijn on 20-5-2015.
  */
-@RunWith(Parameterized.class)
 public class LanguageParserTest {
 
-	private Rule rule;
-	private String input;
-	private boolean match;
+	private LanguageParser parser;
 
-	public LanguageParserTest(Rule rule, String input, boolean match) {
-		this.rule = rule;
-		this.input = input;
-		this.match = match;
-	}
-
-
-	@Parameterized.Parameters
-	public static Collection<Object[]> data() {
-		LanguageParser parser = Parboiled.createParser(LanguageParser.class);
-
-		Object[][] testData = new Object[][] {
-				{ parser.Identifier(), "abOis9d", true },
-				{ parser.Identifier(), "92", false },
-				{ parser.StringLiteral(), "\"dinge niu!h\"", true },
-				{ parser.StringLiteral(), "\"\"", true },
-				{ parser.StringLiteral(), "\"dsa", false },
-				{ parser.Digit(), "9", true },
-				{ parser.Digit(), "abS", false },
-				{ parser.FloatLiteral(), "312.412", true },
-				{ parser.FloatLiteral(), "2", false },
-				{ parser.FloatLiteral(), "s", false },
-				{ parser.IntLiteral(), "4", true },
-				{ parser.IntLiteral(), "415", true },
-				{ parser.Params(), "()", true },
-				{ parser.Params(), "(test)", true },
-				{ parser.Params(), "(test,dingen)", true },
-				{ parser.Params(), "(test, dingen, nogMeer)", true },
-				{ parser.Params(), "(test, 5.0, 3, \"Test\")", true },
-				{ parser.Process(), "test()", true },
-				{ parser.Pipe(), "test()|test2()", true}
-		};
-
-		return Arrays.asList(testData);
+	@Before
+	public void setUp() throws Exception {
+		parser = Parboiled.createParser(LanguageParser.class);
 	}
 
 	@Test
-	public void test() throws Exception {
-		assertEquals(match, new BasicParseRunner<>(rule).run(input).matched);
+	public void testProcess() throws Exception {
+		BasicParseRunner runner = new BasicParseRunner(parser.Process());
+		ParsingResult result = runner.run("test(3, 4, 5)");
+		ProcessInfo info = (ProcessInfo) result.valueStack.pop();
+
+		assertEquals("test", info.getIdentifier().getName());
+		assertArrayEquals(new Object[]{3, 4, 5}, info.getParameters());
+	}
+
+	@Test
+	public void testPipe() throws Exception {
+		BasicParseRunner runner = new BasicParseRunner(parser.Pipe());
+		ParsingResult result = runner.run("test()|test2(\"aap\", sjon, 2.0)");
+		ProcessInfo info = (ProcessInfo) result.valueStack.pop();
+
+		assertEquals("test2", info.getIdentifier().getName());
+		assertArrayEquals(
+				new Object[]{"aap", new Identifier<>("sjon"), 2.0f},
+				info.getParameters()
+		);
+
+		info = (ProcessInfo) result.valueStack.pop();
+
+		assertEquals("test", info.getIdentifier().getName());
+		assertArrayEquals(
+				new Object[]{},
+				info.getParameters()
+		);
 	}
 }
